Relocate throwables
Relocated the throwables to their own package for better file
management.
