--- src/main/game/breakpoint/DeathFail.java
@@ -1,10 +1,0 @@
-package main.game.breakpoint;
-
-public class DeathFail
-		extends GameEnd
-{
-	public DeathFail()
-	{
-		super("The players ran out of lives.");
-	}
-}
