--- src/main/game/breakpoint/DeckEmpty.java
@@ -1,9 +1,0 @@
-package main.game.breakpoint;
-
-public class DeckEmpty extends GameEnd
-{
-	public DeckEmpty()
-	{
-		super("The deck is empty.");
-	}
-}
