--- src/main/game/breakpoint/FullScore.java
@@ -1,9 +1,0 @@
-package main.game.breakpoint;
-
-public class FullScore extends GameEnd
-{
-	public FullScore()
-	{
-		super("The game ended with the maximum number of points available.");
-	}
-}
