--- src/main/game/breakpoint/GameEnd.java
@@ -1,9 +1,0 @@
-package main.game.breakpoint;
-
-public class GameEnd extends Throwable
-{
-	public GameEnd(String message)
-	{
-		super(message);
-	}
-}
