+++ src/main/game/breakpoints/DeathFail.java
@@ -1,0 +1,10 @@
+package main.game.breakpoints;
+
+public class DeathFail
+		extends GameEnd
+{
+	public DeathFail()
+	{
+		super("The players ran out of lives.");
+	}
+}
