+++ src/main/game/breakpoints/DeckEmpty.java
@@ -1,0 +1,9 @@
+package main.game.breakpoints;
+
+public class DeckEmpty extends GameEnd
+{
+	public DeckEmpty()
+	{
+		super("The deck is empty.");
+	}
+}
