+++ src/main/game/breakpoints/GameEnd.java
@@ -1,0 +1,9 @@
+package main.game.breakpoints;
+
+public class GameEnd extends Throwable
+{
+	public GameEnd(String message)
+	{
+		super(message);
+	}
+}
