+++ src/main/game/breakpoints/MaxScore.java
@@ -1,0 +1,10 @@
+package main.game.breakpoints;
+
+public class MaxScore
+		extends GameEnd
+{
+	public MaxScore()
+	{
+		super("The game ended with the maximum number of points available.");
+	}
+}
